+++ src/main/java/me/moomoo/anarchyexploitfixes/Main.java
@@ -1,5 +1,10 @@
 package me.moomoo.anarchyexploitfixes;
 
+import me.moomoo.anarchyexploitfixes.misc.Chat;
+import me.moomoo.anarchyexploitfixes.misc.Expansions;
+import me.moomoo.anarchyexploitfixes.misc.Kicks;
+import me.moomoo.anarchyexploitfixes.patches.*;
+import me.moomoo.anarchyexploitfixes.prevention.*;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Chunk;
