+++ src/main/java/me/moomoo/anarchyexploitfixes/misc/Chat.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.misc;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.World;
@@ -14,6 +15,7 @@ public class Chat implements Listener {
     private final static String REGEX_CHECK = "[-a-zA-Z0-9@:%._+~#=]{1,256}\\.[a-zA-Z()]{1,6}\\b([-a-zA-Z()@:%_+.~#?&/=]*)";
     HashSet<String> antispam = new HashSet<>();
     HashSet<String> words = new HashSet<>();
+
     // shit skidded from john200410, just added another packet to check. msg me if you want removed
     @EventHandler
     private void onChat(AsyncPlayerChatEvent event) {
