+++ src/main/java/me/moomoo/anarchyexploitfixes/misc/Expansions.java
@@ -1,57 +1,58 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.misc;
 
 import me.clip.placeholderapi.expansion.PlaceholderExpansion;
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 
 public class Expansions extends PlaceholderExpansion {
     private final Main plugin;
 
-    public Expansions(Main plugin){
+    public Expansions(Main plugin) {
         this.plugin = plugin;
     }
 
     @Override
-    public boolean persist(){
+    public boolean persist() {
         return true;
     }
 
     @Override
-    public boolean canRegister(){
+    public boolean canRegister() {
         return true;
     }
 
     @Override
-    public @NotNull String getAuthor(){
+    public @NotNull String getAuthor() {
         return "moo";
     }
 
     @Override
-    public @NotNull String getIdentifier(){
+    public @NotNull String getIdentifier() {
         return "chunks";
     }
 
     @Override
-    public @NotNull String getVersion(){
+    public @NotNull String getVersion() {
         return plugin.getDescription().getVersion();
     }
 
     @Override
-    public String onPlaceholderRequest(Player player, @NotNull String identifier){
-        if (player == null){
+    public String onPlaceholderRequest(Player player, @NotNull String identifier) {
+        if (player == null) {
             return "";
         }
 
-        if (identifier.equals("new")){
-            if(Main.new_total[0] > 50){
+        if (identifier.equals("new")) {
+            if (Main.new_total[0] > 50) {
                 return Main.new_total[0].toString();
             } else {
                 return "";
             }
         }
 
-        if (identifier.equals("old")){
-            if(Main.new_total[0] > 50){
+        if (identifier.equals("old")) {
+            if (Main.new_total[0] > 50) {
                 return Main.old_total[0].toString();
             } else {
                 return "";
