+++ src/main/java/me/moomoo/anarchyexploitfixes/misc/Kicks.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.misc;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.ChatColor;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
