+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/Boats.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -14,6 +15,7 @@ import org.bukkit.inventory.PlayerInventory;
 
 public class Boats implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onMove(PlayerMoveEvent evt) {
         if (config.getBoolean("BoatflyPatch")) {
