+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/CrashExploits.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -7,6 +8,7 @@ import org.bukkit.event.entity.EntityTeleportEvent;
 
 public class CrashExploits implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onEntityTeleportEvent(EntityTeleportEvent evt) {
         if (config.getBoolean("PreventEndGatewayCrashExploit")) {
