+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/DupeFixes.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.Material;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.ChestedHorse;
@@ -14,6 +15,7 @@ import org.bukkit.scheduler.BukkitRunnable;
 
 public class DupeFixes implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onEntityPortalEvent(EntityPortalEvent evt) {
         if (config.getBoolean("PreventDonkeysFromGoingThroughPortals")) {
@@ -31,6 +33,7 @@ public class DupeFixes implements Listener {
             }
         }
     }
+
     @EventHandler
     private void onEntityInteract(PlayerInteractAtEntityEvent evt) {
         if (config.getBoolean("DisableChestsOnDonkeys")) {
