+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/Elytra.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.*;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Entity;
@@ -18,6 +19,7 @@ import java.util.HashSet;
 public class Elytra implements Listener {
     public HashSet<String> playersinnewchunks = new HashSet<>();
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onMove(PlayerMoveEvent evt) {
         if (config.getBoolean("DisableAllElytras")) {
@@ -143,6 +145,7 @@ public class Elytra implements Listener {
             }
         }
     }
+
     @EventHandler
     public void onChunkLoad(ChunkLoadEvent evt) {
         if (evt.isNewChunk()) {
