+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/EndPortalPatch.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.Material;
 import org.bukkit.block.BlockFace;
 import org.bukkit.configuration.file.FileConfiguration;
@@ -9,6 +10,7 @@ import org.bukkit.event.player.PlayerBucketEmptyEvent;
 
 public class EndPortalPatch implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onPlayerBucketEvent(PlayerBucketEmptyEvent evt) {
         if (config.getBoolean("PreventDestroyingEndPortals")) {
