+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/GodMode.java
@@ -1,6 +1,7 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
 import io.netty.channel.*;
+import me.moomoo.anarchyexploitfixes.Main;
 import net.minecraft.server.v1_12_R1.PacketPlayInFlying;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -15,6 +16,7 @@ import org.bukkit.event.player.PlayerQuitEvent;
 
 public class GodMode implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     public void onJoin(PlayerJoinEvent event) {
         if (config.getBoolean("PreventGodMode")) {
@@ -23,6 +25,7 @@ public class GodMode implements Listener {
             Main.getPlugin(Main.class).b.put(event.getPlayer(), 0);
         }
     }
+
     @EventHandler
     public void onLeave(PlayerQuitEvent event) {
         if (config.getBoolean("PreventGodMode")) {
@@ -31,6 +34,7 @@ public class GodMode implements Listener {
             Main.getPlugin(Main.class).b.put(event.getPlayer(), 0);
         }
     }
+
     @EventHandler
     public void onMove(PlayerMoveEvent event) {
         if (config.getBoolean("PreventGodMode")) {
@@ -47,6 +51,7 @@ public class GodMode implements Listener {
             }
         }
     }
+
     static void removePlayer(Player player) {
         Channel channel = ((CraftPlayer) player).getHandle().playerConnection.networkManager.channel;
         channel.eventLoop().submit(() -> {
@@ -54,6 +59,7 @@ public class GodMode implements Listener {
             return null;
         });
     }
+
     static void injectPlayer(Player player) {
         ChannelDuplexHandler channelDuplexHandler = new ChannelDuplexHandler() {
