+++ src/main/java/me/moomoo/anarchyexploitfixes/patches/Illegals.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.patches;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.ChatColor;
 import org.bukkit.Chunk;
 import org.bukkit.Material;
@@ -75,6 +76,7 @@ public class Illegals implements Listener {
             });
         }
     }
+
     @EventHandler
     private void onBlockDropItemEvent(PlayerDropItemEvent evt) {
         if (config.getBoolean("DEBUG")) {
@@ -105,6 +107,7 @@ public class Illegals implements Listener {
             });
         }
     }
+
     @EventHandler
     private void onChunkLoadEvent(ChunkLoadEvent evt) {
         if (config.getBoolean("RemoveALLIllegalBlocksOnCHUNKLOAD")) {
@@ -139,6 +142,7 @@ public class Illegals implements Listener {
             }
         }
     }
+
     @EventHandler
     private void onInventoryInteractEvent(InventoryInteractEvent evt) {
         if (config.getBoolean("DEBUG")) {
@@ -162,6 +166,7 @@ public class Illegals implements Listener {
             });
         }
     }
+
     @EventHandler
     private void onInventoryClick(InventoryClickEvent evt) {
         if (config.getBoolean("RemoveIllegalBlocksOnClick")) {
@@ -192,6 +197,7 @@ public class Illegals implements Listener {
             }
         }
     }
+
     @EventHandler
     private void onInventoryPickup(InventoryPickupItemEvent evt) {
         if (config.getBoolean("RemoveIllegalBlocksOnPickup")) {
@@ -213,6 +219,7 @@ public class Illegals implements Listener {
             });
         }
     }
+
     @EventHandler
     private void onInventoryMove(InventoryMoveItemEvent evt) {
         if (config.getBoolean("RemoveIllegalBlocksOnMove")) {
@@ -233,6 +240,7 @@ public class Illegals implements Listener {
             }
         }
     }
+
     @EventHandler
     private void onBlockPlace(BlockPlaceEvent evt) {
         if (config.getBoolean("PreventPlacingIllegalBlocks")) {
