+++ src/main/java/me/moomoo/anarchyexploitfixes/prevention/Bedrock.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.prevention;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.Chunk;
 import org.bukkit.Material;
 import org.bukkit.configuration.file.FileConfiguration;
@@ -9,6 +10,7 @@ import org.bukkit.event.world.ChunkLoadEvent;
 
 public class Bedrock implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onChunkLoadEvent(ChunkLoadEvent evt) {
         if (config.getBoolean("FillInBedrockFloor")) {
