+++ src/main/java/me/moomoo/anarchyexploitfixes/prevention/DisableWithers.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.prevention;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -7,6 +8,7 @@ import org.bukkit.event.entity.ProjectileLaunchEvent;
 
 public class DisableWithers implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     public void onProjectileLaunch(ProjectileLaunchEvent evt) {
         String entity = evt.getEntity().toString(); // The entity
