+++ src/main/java/me/moomoo/anarchyexploitfixes/prevention/Joinmessages.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.prevention;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -8,6 +9,7 @@ import org.bukkit.event.player.PlayerQuitEvent;
 
 public class Joinmessages implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onPlayerJoinEvent(PlayerJoinEvent evt) {
         if (config.getBoolean("GrayJoinLeaveMessages")) {
