+++ src/main/java/me/moomoo/anarchyexploitfixes/prevention/LagExploits.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.prevention;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
@@ -8,6 +9,7 @@ import org.bukkit.event.block.BlockPhysicsEvent;
 
 public class LagExploits implements Listener {
     FileConfiguration config = Main.getPlugin(Main.class).getConfig();
+
     @EventHandler
     private void onEntityChange(BlockPhysicsEvent evt) {
         double tps_double = Bukkit.getServer().getTPS()[0];
@@ -15,5 +17,4 @@ public class LagExploits implements Listener {
             evt.setCancelled(true);
         }
     }
-
 }
