+++ src/main/java/me/moomoo/anarchyexploitfixes/prevention/NetherRoof.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.prevention;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.*;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
