+++ src/main/java/me/moomoo/anarchyexploitfixes/prevention/Redstone.java
@@ -1,5 +1,6 @@
-package me.moomoo.anarchyexploitfixes;
+package me.moomoo.anarchyexploitfixes.prevention;
 
+import me.moomoo.anarchyexploitfixes.Main;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
@@ -30,6 +31,7 @@ public class Redstone implements Listener {
             evt.setNewCurrent(newcurrent);
         }
     }
+
     @EventHandler
     private void onExplode(EntityExplodeEvent evt) {
         double tps = Bukkit.getServer().getTPS()[0];
